requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.4.2
python-multipart==0.0.6
gunicorn==21.2.0
mysqlclient==2.2.0
sqlalchemy==2.0.23
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
requests==2.31.0
pymysql==1.1.0
cryptography==41.0.5
psycopg2-binary==2.9.9
pytest==7.4.3